fix: scroll issue on theme selector
